Moved CopyAndReleaseBufferDataCommand and CopyAndReleaseBufferDataCommand out of ImageData and BufferData respectively
