+++ include/vsg/traversals/CompileTraversal.h
@@ -26,6 +26,9 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/vk/Fence.h>
 #include <vsg/vk/Command.h>
 
+#include <vsg/vk/BufferData.h>
+#include <vsg/vk/ImageData.h>
+
 namespace vsg
 {
     struct BufferPreferences
@@ -68,6 +71,42 @@ namespace vsg
     };
 
 
+    class CopyAndReleaseBufferDataCommand : public Command
+    {
+    public:
+
+        CopyAndReleaseBufferDataCommand(BufferData src, BufferData dest):
+            source(src),
+            destination(dest) {}
+
+        BufferData source;
+        BufferData destination;
+
+        void dispatch(CommandBuffer& commandBuffer) const override;
+
+    protected:
+        virtual ~CopyAndReleaseBufferDataCommand();
+    };
+
+    class CopyAndReleaseImageDataCommand : public Command
+    {
+    public:
+
+        CopyAndReleaseImageDataCommand(BufferData src, ImageData dest, uint32_t numMipMapLevels):
+            source(src),
+            destination(dest),
+            mipLevels(numMipMapLevels) {}
+
+        void dispatch(CommandBuffer& commandBuffer) const override;
+
+        BufferData source;
+        ImageData destination;
+        uint32_t mipLevels = 1;
+
+    protected:
+        virtual ~CopyAndReleaseImageDataCommand();
+    };
+
     class Context
     {
     public:
@@ -93,6 +132,8 @@ namespace vsg
         ref_ptr<CommandBuffer> commandBuffer;
         ref_ptr<Fence> fence;
 
+        std::vector<ref_ptr<CopyAndReleaseBufferDataCommand>> copyBufferDataCommands;
+        std::vector<ref_ptr<CopyAndReleaseImageDataCommand>> copyImageDataCommands;
         std::vector<ref_ptr<Command>> commands;
 
         void dispatchCommands();
